- combine more common functioanlity (all of it!) - skeleton class for JDK client
