+++ solr/solrj/src/java/org/apache/solr/client/solrj/impl/Http2SolrClient.java
@@ -821,10 +821,6 @@ public class Http2SolrClient extends HttpSolrClientBase {
         .collect(Collectors.joining(", "));
   }
 
-  protected RequestWriter getRequestWriter() {
-    return requestWriter;
-  }
-
   /**
    * An Http2SolrClient that doesn't close or cleanup any resources
    *
