+++ solr/solrj/src/java/org/apache/solr/client/solrj/impl/LBHttp2SolrClient.java
@@ -1,38 +1,9 @@
 package org.apache.solr.client.solrj.impl;
 
-import org.apache.solr.client.solrj.ResponseParser;
-import org.apache.solr.client.solrj.SolrClient;
-import org.apache.solr.client.solrj.request.RequestWriter;
-
-import java.util.Set;
-
-public class LBHttp2SolrClient extends LBHttpSolrClientBase {
-    protected final Http2SolrClient solrClient;
+public class LBHttp2SolrClient extends LBHttpSolrClientBase<Http2SolrClient> {
 
     protected LBHttp2SolrClient(Builder builder) {
         super(builder);
-        this.solrClient = builder.http2SolrClient;
-        this.aliveCheckIntervalMillis = builder.aliveCheckIntervalMillis;
-        this.defaultCollection = builder.defaultCollection;
-    }
-
-    @Override
-    protected SolrClient getClient(Endpoint endpoint) {
-        return solrClient;
-    }
-
-    @Override
-    public ResponseParser getParser() {
-        return solrClient.getParser();
-    }
-
-    @Override
-    public RequestWriter getRequestWriter() {
-        return solrClient.getRequestWriter();
-    }
-
-    public Set<String> getUrlParamNames() {
-        return solrClient.getUrlParamNames();
     }
 
     public static class Builder extends LBHttpSolrClientBuilderBase<LBHttp2SolrClient, LBHttp2SolrClient.Builder, Http2SolrClient> {
