+++ solr/solrj/src/java/org/apache/solr/client/solrj/impl/LBHttpSolrClientBase.java
@@ -23,12 +23,16 @@ import java.net.ConnectException;
 import java.net.SocketException;
 import java.net.SocketTimeoutException;
 import java.util.Arrays;
+import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 
+import org.apache.solr.client.solrj.ResponseParser;
+import org.apache.solr.client.solrj.SolrClient;
 import org.apache.solr.client.solrj.SolrServerException;
 import org.apache.solr.client.solrj.request.IsUpdateRequest;
+import org.apache.solr.client.solrj.request.RequestWriter;
 import org.apache.solr.common.SolrException;
 import org.apache.solr.common.util.NamedList;
 import org.slf4j.MDC;
@@ -92,14 +96,36 @@ import org.slf4j.MDC;
  *
  * @since solr 8.0
  */
-public abstract class LBHttpSolrClientBase extends LBSolrClient {
+public abstract class LBHttpSolrClientBase<C extends HttpSolrClientBase> extends LBSolrClient {
 
+  protected final C solrClient;
 
+  @SuppressWarnings("unchecked")
   protected LBHttpSolrClientBase(LBHttpSolrClientBuilderBase<?,?,?> builder) {
     super(Arrays.asList(builder.solrEndpoints));
+    this.solrClient = (C) builder.solrClient;
+    this.aliveCheckIntervalMillis = builder.aliveCheckIntervalMillis;
+    this.defaultCollection = builder.defaultCollection;
   }
 
-  
+  @Override
+  protected SolrClient getClient(Endpoint endpoint) {
+    return solrClient;
+  }
+
+  @Override
+  public ResponseParser getParser() {
+    return solrClient.getParser();
+  }
+
+  @Override
+  public RequestWriter getRequestWriter() {
+    return solrClient.getRequestWriter();
+  }
+
+  public Set<String> getUrlParamNames() {
+    return solrClient.getUrlParamNames();
+  }
 
   /**
    * Execute an asynchronous request against one or more hosts for a given collection. The passed-in
